+++ arangod/Cluster/ResultT.h
@@ -164,7 +164,7 @@ class ResultT {
   bool ok() const { return _result.ok(); }
   bool fail() const { return _result.fail(); }
   bool is(uint64_t code) { return _result.is(code); }
-  uint64_t errorNumber() const { return _result.errorNumber(); }
+  int errorNumber() const { return _result.errorNumber(); }
   std::string errorMessage() const { return _result.errorMessage(); }
 
   // access methods
